+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - the `search`, `export`, and `modify` commands now also support sorting and limiting options to be
   passed onto the `list` command (along with the already existing filtering options)
 
+### Changed
+- when using the TUI you may now use the `:show <label>` command to jump to the
+  specified label (#126,!116)
+
 ### Fixed
 - unblocked the vertical scrollbar of the TUI's list view from the scroll offset